restructure conduit folder
