+++ src/analysis/retail/monk/shared/hero/ConduitOfTheCelestials/constants.ts
@@ -1,0 +1,3 @@
+export const CELESTIAL_CONDUIT_MAX_DURATION = 4000;
+export const CELESTIAL_CONDUIT_INCREASE_PER_TARGET = 0.06;
+export const CELESTIAL_CONDUIT_MAX_INCREASE = 0.3;
